+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `FDNetwork.number_of_supports()` to count number of supported nodes in a network.
+- Added `network_validate` to check the validity of a network before equilibrium calculations.
+
 ### Changed
 
 ### Removed